update test common and add test_downloads for functions moved from common to downloads
