+++ tests/test_common.py
@@ -6,7 +6,6 @@ from datetime import timezone
 import tempfile
 import shutil
 
-from coastseg import exceptions
 from coastseg import common
 from coastseg import file_utilities
 
@@ -17,9 +16,6 @@ from shapely import geometry
 import pandas as pd
 import pytest
 from unittest.mock import patch, call
-import pytest
-from coastseg import common
-from typing import Dict, List, Union
 
 def test_empty_merged_timeseries_gdf():
     # Create an empty GeoDataFrame with the necessary structure
@@ -73,7 +69,7 @@ def test_get_seaward_points_gdf():
             ],
         },
         crs = 4326
-    ) 
+    )  # type: ignore
 
     # Call the function to get the seaward points GeoDataFrame
     seaward_points_gdf = common.get_seaward_points_gdf(transects)
@@ -100,7 +96,7 @@ def test_get_seaward_points_gdf_diff_crs():
             ],
         },
         crs = 3857
-    ) 
+    ) # type: ignore
 
     # Call the function to get the seaward points GeoDataFrame
     seaward_points_gdf = common.get_seaward_points_gdf(transects)
@@ -117,58 +113,6 @@ def test_get_seaward_points_gdf_diff_crs():
     #         38.11274239609162)
  
 
-
-def test_authenticate_and_initialize_success():
-    with patch('coastseg.common.ee.Authenticate') as mock_authenticate, \
-         patch('coastseg.common.ee.Initialize') as mock_initialize:
-        
-        # Mock successful initialization
-        mock_initialize.return_value = None
-        
-        common.authenticate_and_initialize(print_mode=True, force=False, auth_args={}, kwargs={})
-
-        mock_authenticate.assert_called_once() # this will call once becase ee.credentials is None
-        mock_initialize.assert_called_once()
-
-def test_authenticate_and_initialize_force_auth():
-    with patch('coastseg.common.ee.Authenticate') as mock_authenticate, \
-         patch('coastseg.common.ee.Initialize') as mock_initialize:
-        
-        # Mock successful initialization
-        mock_initialize.return_value = None
-        
-        common.authenticate_and_initialize(print_mode=True, force=True, auth_args={}, kwargs={})
-
-        mock_authenticate.assert_called_once_with(force=True)
-        mock_initialize.assert_called_once()
-
-def test_authenticate_and_initialize_retry():
-    with patch('coastseg.common.ee.Authenticate') as mock_authenticate, \
-         patch('coastseg.common.ee.Initialize') as mock_initialize:
-        
-        # Mock an exception on first initialize, then success
-        mock_initialize.side_effect = [Exception("Credentials file not found"), None]
-
-        common.authenticate_and_initialize(print_mode=True, force=False, auth_args={}, kwargs={})
-
-        assert mock_authenticate.call_count == 2
-        assert mock_initialize.call_count == 2
-
-def test_authenticate_and_initialize_max_attempts():
-    with patch('coastseg.common.ee.Authenticate') as mock_authenticate, \
-         patch('coastseg.common.ee.Initialize') as mock_initialize:
-        
-        # Mock an exception for all initialize attempts
-        mock_initialize.side_effect = Exception("Credentials file not found")
-        
-        with pytest.raises(Exception) as excinfo:
-            common.authenticate_and_initialize(print_mode=True, force=False, auth_args={}, kwargs={})
-        
-        assert "Failed to initialize Google Earth Engine after 2 attempts" in str(excinfo.value)
-        assert mock_authenticate.call_count == 2
-        assert mock_initialize.call_count == 2
-
-
 def test_order_linestrings_gdf_empty():
     gdf = gpd.GeoDataFrame({'geometry': []})
     result = common.order_linestrings_gdf(gdf)
@@ -176,7 +120,7 @@ def test_order_linestrings_gdf_empty():
 
 def test_order_linestrings_gdf_single_linestring():
     points = np.array([[0, 0], [1, 1]])
-    gdf = gpd.GeoDataFrame({'geometry': [LineString(points)],'date': ['2023-01-01']}, crs='epsg:4326')
+    gdf = gpd.GeoDataFrame({'geometry': [LineString(points)],'date': ['2023-01-01']}, crs='epsg:4326') #type: ignore
     result = common.order_linestrings_gdf(gdf)
     expected_line = common.create_complete_line_string(points)
     assert len(result) == 1, "Expected one linestring in the result"
